Add JMeter JSON collector.
