+++ components/collector/tests/source_collectors/jmeter_json/test_slow_transactions.py
@@ -1,40 +1,12 @@
 """Unit tests for the JMeter JSON slow transactions collector."""
 
-from ..source_collector_test_case import SourceCollectorTestCase
+from .base import JMeterJSONTestCase
 
 
-class JMeterJSONSlowTransactionsTest(SourceCollectorTestCase):
+class JMeterJSONSlowTransactionsTest(JMeterJSONTestCase):
     """Unit tests for the JMeter JSON slow transactions collector."""
 
     METRIC_TYPE = "slow_transactions"
-    SOURCE_TYPE = "jmeter_json"
-    API1 = "/api/foo"
-    API2 = "/api/bar"
-    JMETER_JSON = json = dict(
-        transaction1=dict(
-            transaction=API1,
-            sampleCount=123,
-            errorCount=2,
-            errorPct=2 / 123,
-            meanResTime=110,
-            medianResTime=120,
-            minResTime=50.0,
-            maxResTime=250.0000004,
-            pct1ResTime=115.0,
-        ),
-        transaction2=dict(
-            transaction=API2,
-            sampleCount=125,
-            errorCount=4,
-            errorPct=4 / 125,
-            meanResTime=110.56,
-            medianResTime=130,
-            minResTime=40.0,
-            maxResTime=2500.03223,
-            pct1ResTime=120.0,
-        ),
-        Total={},  # Total should be ignored
-    )
 
     def setUp(self):
         """Extend to set the expected entities."""
